refactor: move transaction input types from postgres adapter to pkg/transaction :hammer:
